feat: Refactor Handlers & Plugins
Add Bun workspaces to leverage plugin & dependency installation.
Add plugin system.
Update Dockerfile and supervisord configuration to reflect changes.
