+++ README.md
@@ -9,6 +9,7 @@ _mounts/terminologies/data/icd10cm-tabular-2023.xml_
 - `id`: the code-system identifier
 - `type`: upload-terminology
 - `source`: the mounted code-system source file
+- `plugin`: the plugin module to use for processing the code-system (optional)
 
 _mounts/configs/bonfhir-hapi.yml_
 
@@ -54,19 +55,19 @@ mvn -Pjetty jetty:run
 
 ## Integrating custom terminologies
 
-Add a new `upload-terminology-plugin` task to the `bonfhir-hapi.yml` configuration file, specifying the `id` of the terminology, the `source` file, and the `plugin` to use for processing the terminology.
+Add a new `upload-terminology` task to the `bonfhir-hapi.yml` configuration file, specifying the `id` of the terminology, the `source` file, and the `plugin` pointing to the module to use for processing the terminology.
 
 ```yml
-- type: upload-terminology-plugin
+- type: upload-terminology
     id: http://www.nlm.nih.gov/research/umls/rxnorm
     source: RxNorm_full_03042024.zip
     plugin: rxnorm
 ```
 
-Create a new plugin package for your plugin. You can copy the rxnorm plugin as base.
+Create a new plugin package for your plugin. You can copy the rxnorm or CPT/HCPCS plugin as base.
 
 ```yml
-- type: upload-terminology-plugin
+- type: upload-terminology
     id: https://www.cms.gov/Medicare/Coding/HCPCSReleaseCodeSets
     source: 2024_DHS_Code_List_Addendum_03_01_2024.zip
     plugin: cpt-hcpcs
