+++ index.ts
@@ -1,0 +1,1 @@
+import "terminology-server-setup/bootstrap";
