+++ packages/server-setup/configs.ts
@@ -4,10 +4,7 @@ export const FHIR_VERSIONS = ["R3", "R4", "R4B", "R5"] as const;
 export type FHIRVersion = (typeof FHIR_VERSIONS)[number];
 
 export type ConfigTaskEntry = {
-  type:
-    | "upload-definitions"
-    | "upload-terminology"
-    | "upload-terminology-plugin";
+  type: "upload-definitions" | "upload-terminology";
   id: string;
   source: string;
   plugin?: string;
