+++ packages/server-setup/package.json
@@ -1,6 +1,5 @@
 {
-  "name": "terminology-server-scripts",
-  "module": "server-setup",
+  "name": "terminology-server-setup",
   "type": "module",
   "devDependencies": {
     "@types/bun": "latest",
