+++ supervisord.conf
@@ -22,8 +22,8 @@ redirect_stderr=true
 
 [program:server-bootstrap]
 user=bonfhir
-directory=/bonfhir/server-setup
-command=/bonfhir/server-setup/bootstrap.ts
+directory=/bonfhir
+command=bun start
 stdout_logfile=/dev/fd/1
 stdout_logfile_maxbytes=0
 redirect_stderr=true
